refactor: MinIoTest local qr 코드 저장 코드 삭제
