--- qrust-api/src/main/java/com/qrust/common/exception/error/ErrorCode.java
@@ -1,30 +1,0 @@
-package com.qrust.common.exception.error;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum ErrorCode {
-    // TODO: 테스트용 (추후 제거)
-    TEST_ERROR(10000, HttpStatus.BAD_REQUEST, "테스트 에러입니다."),
-
-    // 400 Bad Request
-    INVALID_INPUT_VALUE(40000, HttpStatus.BAD_REQUEST, "유효하지 않은 입력 값입니다."),
-    INVALID_TYPE_VALUE(40001, HttpStatus.BAD_REQUEST, "요청 파라미터 타입이 올바르지 않습니다."),
-    HTTP_MESSAGE_NOT_READABLE(40002, HttpStatus.BAD_REQUEST, "요청 본문 형식을 읽을 수 없습니다."),
-
-    // 404 Not Found
-    NOT_FOUND_END_POINT(40400, HttpStatus.NOT_FOUND, "존재하지 않는 API입니다."),
-
-    // 405 Method Not Allowed
-    METHOD_NOT_ALLOWED(40500, HttpStatus.METHOD_NOT_ALLOWED, "허용되지 않은 HTTP 메서드입니다."),
-
-    // 500 Internal Server Error
-    INTERNAL_SERVER_ERROR(50000, HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류입니다.");
-
-    private final Integer code;
-    private final HttpStatus httpStatus;
-    private final String message;
-}
