+++ qrust-domain/src/main/java/com/qrust/common/dto/ApiResponse.java
@@ -1,0 +1,28 @@
+package com.qrust.common.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.qrust.common.exception.CustomException;
+import jakarta.validation.constraints.NotNull;
+import org.springframework.http.HttpStatus;
+import org.springframework.lang.Nullable;
+
+public record ApiResponse<T>(
+        @JsonIgnore
+        HttpStatus httpStatus,
+        @NotNull boolean success,
+        @Nullable T data,
+        @Nullable ExceptionDto error
+) {
+
+    public static <T> ApiResponse<T> ok(@Nullable final T data) {
+        return new ApiResponse<>(HttpStatus.OK, true, data, null);
+    }
+
+    public static <T> ApiResponse<T> created(@Nullable final T data) {
+        return new ApiResponse<>(HttpStatus.CREATED, true, data, null);
+    }
+
+    public static <T> ApiResponse<T> fail(final CustomException e) {
+        return new ApiResponse<>(e.getErrorCode().getHttpStatus(), false, null, ExceptionDto.of(e.getErrorCode()));
+    }
+}
