+++ qrust-domain/src/main/java/com/qrust/common/dto/ApiResponseAdvice.java
@@ -1,0 +1,30 @@
+package com.qrust.common.dto;
+
+import com.qrust.common.dto.ApiResponse;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+@RestControllerAdvice
+public class ApiResponseAdvice implements ResponseBodyAdvice {
+
+    @Override
+    public boolean supports(MethodParameter returnType, Class converterType) {
+        return true;
+    }
+
+    @Override
+    public Object beforeBodyWrite(Object body, MethodParameter returnType, MediaType selectedContentType,
+                                  Class selectedConverterType, ServerHttpRequest request, ServerHttpResponse response) {
+        if (body instanceof ApiResponse<?>) {
+            HttpStatus status = ((ApiResponse<?>) body).httpStatus();
+            response.setStatusCode(status);
+        }
+
+        return body;
+    }
+}
