+++ qrust-domain/src/main/java/com/qrust/user/domain/entity/User.java
@@ -1,6 +1,6 @@
 package com.qrust.user.domain.entity;
 
-import com.qrust.infrastructure.jpa.shared.BaseEntity;
+import com.qrust.common.infrastructure.jpa.shared.BaseEntity;
 import com.qrust.user.domain.entity.vo.Gender;
 import com.qrust.user.domain.entity.vo.LoginType;
 import com.qrust.user.domain.entity.vo.UserRole;
